refactor: move creator components to separate folder
