+++ src/components/CreatorComponents/index.js
@@ -1,0 +1,2 @@
+export { default as EducationInput } from "./EducationInput";
+export { default as WorkExperienceInput } from "./WorkExperienceInput";
