+++ src/components/FormComponents/index.js
@@ -1,6 +1,5 @@
-export { default as EducationInput } from "./EducationInput";
 export { default as FieldsInput } from "./FieldsInput";
 export { default as FormField } from "./FormField";
-export { default as WorkExperienceInput } from "./WorkExperienceInput";
+export { default as MoveControls } from "./MoveControls";
 
 export * from "./styledComponents";
